requirements.txt
# Dependências principais
crewai>=0.70.0
streamlit>=1.28.0
pyyaml>=6.0
pathlib2>=2.3.7

# Processamento de documentos
PyPDF2>=3.0.0
python-docx>=0.8.11

# Análise e manipulação de dados
pandas>=2.0.0
numpy>=1.24.0

# Interface web e visualização
plotly>=5.15.0
matplotlib>=3.7.0

# Utilitários
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# Processamento de texto
langchain>=0.1.0
langchain-community>=0.0.20

# LLM e IA
# watsonx-ai>=1.0.0

# Desenvolvimento e testes (opcionais)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Logs e monitoramento
loguru>=0.7.0 